Some minor changes (#8)
* some minor changes

* removed not finished file
