+++ artifacts/cw_ibc_example.wasm
